+++ packages/cli/CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to the Zowe CLI package will be documented in this file.
 
 - BugFix: Update zos-files copy dsclp system tests to include large mock files. 
 
+## `7.22.0`
+
+- Enhancement: Hid the progress bar if `CI` environment variable is set, or if `FORCE_COLOR` environment variable is set to `0`. [#1845](https://github.com/zowe/zowe-cli/issues/1845)
+
 ## `7.21.2`
 
 - BugFix: Correct extra character being displayed at the end of lines when issuing `zowe files compare` on Windows. [#1992](https://github.com/zowe/zowe-cli/issues/1992)